[Ocean Adam]should_return_204_code_when_set_user_to_orders controller test

+++ src/main/resources/db/migration/V1__Create_ORDERS_Table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE `orders` (
   `id`             INT       NOT NULL AUTO_INCREMENT PRIMARY KEY,
-  `user_id`        BIGINT,
+  `user_id`        INT,
   `type`           VARCHAR(20)  NOT NULL,
   `status`         VARCHAR(20)  NOT NULL,
   `car_id`         VARCHAR(20)  NOT NULL,
